LayoutTests/fast/canvas/webgl/getActiveTest-expected.txt

Test of getActiveAttrib and getActiveUniform

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS context.getError() is 0
PASS context.getActiveUniform(program, 0).name is 'u_modelViewProjMatrix'
PASS context.getActiveUniform(program, 0).type is context.FLOAT_MAT4
PASS context.getActiveUniform(program, 0).size is 1
PASS context.getActiveUniform(program, 1) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS context.getError() is context.INVALID_VALUE
PASS context.getActiveUniform(program, -1) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS context.getError() is context.INVALID_VALUE
PASS context.getActiveUniform(null, 0) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS context.getError() is 0
PASS context.getActiveAttrib(program, 0).name is 'a_normal'
PASS context.getActiveAttrib(program, 0).type is context.FLOAT_VEC3
PASS context.getActiveAttrib(program, 0).size is 1
PASS context.getActiveAttrib(program, 1).name is 'a_vertex'
PASS context.getActiveAttrib(program, 1).type is context.FLOAT_VEC4
PASS context.getActiveAttrib(program, 1).size is 1
PASS context.getActiveAttrib(program, 2) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS context.getError() is context.INVALID_VALUE
PASS context.getActiveAttrib(program, -1) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS context.getError() is context.INVALID_VALUE
PASS context.getActiveAttrib(null, 0) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS context.getError() is 0
PASS context2.getError() is 0
PASS context2.getActiveAttrib(program, 0) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS context2.getError() is 0
PASS context2.getActiveUniform(program, 0) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS context2.getError() is 0
PASS context.getActiveUniform(program, 0) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS context.getError() is 0
PASS context.getActiveAttrib(program, 0) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS context.getError() is 0
PASS successfullyParsed is true

TEST COMPLETE
